+++ elastic_class.py
@@ -650,7 +650,6 @@ class ElasticSearchRepo(ElasticSearch):
         self.repo = repo
         self.repo_dir = repo_dir
         self.repo_dict = {}
-        self.update_repo_status()
 
     def connect(self):