+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [3.0.0] - 2021-10-18
 Breaking Change
 
+### Changed
+- elastic_libs:  Removed non-required \*\*kwargs entries.
+
 
 ## [2.0.3] - 2020-07-10
 ### Fixed