+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Breaking Change
 
 ### Changed
 - elastic_class.ElasticSearchDump.\_\_init\_\_:  Removed update_dump_status call.
-- elastic_class.ElasticSearch.update_status:  Removed is_active check and setting of is_connected attribute.
+- elastic_class.ElasticSearch.update_status elastic_class.ElasticSearchDump.update_dump_status:  Removed is_active check and setting of is_connected attribute.
 - elastic_class.ElasticSearch.\_\_init\_\_:  Removed connection and update_status calls.
 - elastic_libs, elastc_class:  Removed non-required \*\*kwargs entries.
 - Documentation updates