+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Breaking Change
 - elastic_class.ElasticSearch.connect:  Connection method to ElasticSearch server/cluster.
 
 ### Changed
+- elastic_class.ElasticSearchStatus\_\_init\_\_:  Removed update_status2 call.
 - elastic_class.ElasticSearchRepo.\_\_init\_\_:  Removed update_repo_status call.
 - elastic_class.ElasticSearchDump.\_\_init\_\_:  Removed update_dump_status call.
 - elastic_class.ElasticSearchRepo.update_repo_status, elastic_class.ElasticSearch.update_status, elastic_class.ElasticSearchDump.update_dump_status:  Removed is_active check and setting of is_connected attribute.