+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [3.0.1] - 2021-11-24
+
+
 ## [3.0.0] - 2021-10-18
 Breaking Change